* k5-platform.h (store_16_be, store_16_le): Use unsigned int arg, not unsigned short.
git-svn-id: svn://anonsvn.mit.edu/krb5/branches/raeburn-gssapi-cfx@15836 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/include/k5-platform.h
@@ -51,13 +51,13 @@
    etc.)  */
 
 static inline void
-store_16_be (unsigned short val, unsigned char *p)
+store_16_be (unsigned int val, unsigned char *p)
 {
     p[0] = (val >>  8) & 0xff;
     p[1] = (val      ) & 0xff;
 }
 static inline void
-store_16_le (unsigned short val, unsigned char *p)
+store_16_le (unsigned int val, unsigned char *p)
 {
     p[1] = (val >>  8) & 0xff;
     p[0] = (val      ) & 0xff;
